@@ -2,25 +2,18 @@ import { NgModule } from "@angular/core";
 import { ContentManagerModule, SxcRootModule } from "@2sic.com/sxc-angular";
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { FormsModule } from "@angular/forms";
 import { ReactiveFormsModule } from "@angular/forms";
 import { MatIconModule } from "@angular/material/icon";
 import { AppComponent } from "./app.component";
 
-@NgModule({
-  declarations: [AppComponent],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    SxcRootModule,
-    ContentManagerModule,
-    FormsModule,
-    BrowserAnimationsModule,
-    MatIconModule,
-    ReactiveFormsModule,
-  ],
-  providers: [],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [AppComponent],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        SxcRootModule,
+        ContentManagerModule,
+        FormsModule,
+        BrowserAnimationsModule,
+        MatIconModule,
+        ReactiveFormsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
